chore: Ariticle 폴더 정리
