+++ src/pages/article/components/CommentList/CommentListItem/CommentListItem.style.tsx
@@ -19,6 +19,7 @@ export const Comment = styled.article`
 export const ReplyComment = styled(Comment)`
   margin-left: 25px;
 `;
+
 export const CommentContent = styled.p`
   margin-top: 10px;
   font-size: ${(props) => props.theme.fonts.$xs};
